refactor: Cards -> Deck 변경, 기존 Deck 제거
- 불필요한 Wrapping 제거
